pyproject.toml
[project]
name = "news-api-mcp"
version = "0.1.0"
description = "Model Context Protocol (MCP) server for accessing the News API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.1.2",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
news-api-mcp = "news_api_mcp:main"

[[project.authors]]
name = "berlinbra"
email = "BMBpsn@gmail.com"
